update based on PR feedback

+++ Avatar_Micro_Robot/Robot_V2/Power_Board/trunk/closed_loop_control/core/InputCapture.h
@@ -83,5 +83,5 @@ void IC_Deinit(void);
 
 #define MAX_NUM_IC_PINS           9
 
-static volatile float periods[MAX_NUM_IC_PINS];
+static volatile uint16_t periods[MAX_NUM_IC_PINS];
 static volatile int measuredMotorDirection[2];
